# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /genesis/launches/reconcile:
    post:
      operationId: reconcileGenesisLaunch
      tags: [genesis]
      summary: Reconcile and index a completed Genesis launch
      description: Verifies both successful Robinhood Chain transactions and all configured Genesis contract mappings, events, immutable token metadata, and fee-vault wiring before persisting public metadata. Client addresses and metadata are never trusted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenesisLaunchReconcileInput"
      responses:
        "200":
          description: Existing or newly reconciled launch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenesisLaunch"
        "202":
          description: On-chain launch verified; public source indexing is still pending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenesisLaunchIndexingPending"
        "400":
          description: Invalid transaction hashes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Transactions do not describe a valid supported Genesis launch or conflict with an indexed launch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Chain verification unavailable or failed closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /genesis/launches:
    get:
      operationId: listGenesisLaunches
      tags: [genesis]
      summary: List public Genesis launches
      parameters:
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
      responses:
        "200":
          description: Newest indexed public Genesis launches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenesisLaunchList"
        "502":
          description: Chain fee-vault reads unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /genesis/launches/{tokenAddress}:
    get:
      operationId: getGenesisLaunch
      tags: [genesis]
      summary: Get one public Genesis launch
      parameters:
        - $ref: "#/components/parameters/GenesisTokenAddress"
      responses:
        "200":
          description: Public Genesis launch with live creator fee balances
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenesisLaunch"
        "400":
          description: Invalid token address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Genesis launch not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Chain fee-vault reads unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /genesis/creators/{walletAddress}/launches:
    get:
      operationId: listCreatorGenesisLaunches
      tags: [genesis]
      summary: List a creator's public Genesis launches
      parameters:
        - $ref: "#/components/parameters/GenesisWalletAddress"
      responses:
        "200":
          description: Creator dashboard launches with live creator fee balances
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenesisLaunchList"
        "400":
          description: Invalid creator wallet address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Chain fee-vault reads unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    GenesisWalletAddress:
      name: walletAddress
      in: path
      required: true
      schema:
        type: string
        pattern: "^0x[a-fA-F0-9]{40}$"
    GenesisTokenAddress:
      name: tokenAddress
      in: path
      required: true
      schema:
        type: string
        pattern: "^0x[a-fA-F0-9]{40}$"
...
  schemas:
...
    GenesisLaunchReconcileInput:
      type: object
      additionalProperties: false
      required: [version, deploymentTxHash, launchTxHash]
      properties:
        version:
          type: string
          enum: [v1, v2]
        deploymentTxHash: { type: string, pattern: "^0x[a-fA-F0-9]{64}$" }
        launchTxHash: { type: string, pattern: "^0x[a-fA-F0-9]{64}$" }
    GenesisLaunch:
      type: object
      additionalProperties: false
      required: [version, chainId, creatorWalletAddress, genesisDigest, tokenAddress, feeVaultAddress, hookAddress, positionLockAddress, poolId, positionTokenId, tokenName, symbol, logoUri, logoCid, deploymentTxHash, deploymentBlockNumber, deploymentBlockHash, deploymentTimestamp, launchTxHash, launchBlockNumber, launchBlockHash, launchTimestamp, creatorNativeDue, creatorWethDue, creatorGenesisTokenDue, explorer]
      properties:
        version: { type: string, enum: [v1, v2] }
        chainId: { type: integer, enum: [4663] }
        creatorWalletAddress: { type: string, pattern: "^0x[a-fA-F0-9]{40}$" }
        genesisDigest: { type: string, pattern: "^0x[a-fA-F0-9]{64}$" }
        tokenAddress: { type: string, pattern: "^0x[a-fA-F0-9]{40}$" }
        feeVaultAddress: { type: string, pattern: "^0x[a-fA-F0-9]{40}$" }
        hookAddress: { type: string, pattern: "^0x[a-fA-F0-9]{40}$" }
        positionLockAddress: { type: string, pattern: "^0x[a-fA-F0-9]{40}$" }
        poolId: { type: string, pattern: "^0x[a-fA-F0-9]{64}$" }
        positionTokenId: { type: string, pattern: "^[0-9]+$" }
        tokenName: { type: string }
        symbol: { type: string }
        logoUri: { type: string, pattern: "^ipfs://.+" }
        logoCid: { type: string, minLength: 1 }
        deploymentTxHash: { type: string, pattern: "^0x[a-fA-F0-9]{64}$" }
        deploymentBlockNumber: { type: string, pattern: "^[0-9]+$" }
        deploymentBlockHash: { type: string, pattern: "^0x[a-fA-F0-9]{64}$" }
        deploymentTimestamp: { type: string, format: date-time }
        launchTxHash: { type: string, pattern: "^0x[a-fA-F0-9]{64}$" }
        launchBlockNumber: { type: string, pattern: "^[0-9]+$" }
        launchBlockHash: { type: string, pattern: "^0x[a-fA-F0-9]{64}$" }
        launchTimestamp: { type: string, format: date-time }
        creatorNativeDue: { type: string, pattern: "^[0-9]+$" }
        creatorWethDue: { type: string, pattern: "^[0-9]+$" }
        creatorGenesisTokenDue: { type: string, pattern: "^[0-9]+$" }
        explorer:
          type: object
          additionalProperties: false
          required: [token, deploymentTx, launchTx]
          properties:
            token: { type: string, format: uri }
            deploymentTx: { type: string, format: uri }
            launchTx: { type: string, format: uri }
    GenesisLaunchIndexingPending:
      type: object
      additionalProperties: false
      required: [status, onChain, sourcify, scanners]
      properties:
        status: { type: string, enum: [pending_indexing] }
        onChain: { type: string, enum: [verified] }
        sourcify:
          type: object
          additionalProperties: false
          required: [token, hook]
          properties:
            token: { type: string, enum: [pending, exact_match] }
            hook: { type: string, enum: [pending, exact_match] }
        scanners:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [name, token, hook]
            properties:
              name: { type: string }
              token: { type: string, enum: [pending, exact_match] }
              hook: { type: string, enum: [pending, exact_match] }
    GenesisLaunchList:
      type: object
      additionalProperties: false
      required: [items, page, limit, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/GenesisLaunch"
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1, maximum: 100 }
        total: { type: integer, minimum: 0 }
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error